Jens:
Did some divide and conquer on the startServer-function.
This resulted in two new functions, commands and decodeMessage.
Also rewrote the startServer-function and the start-function, so that they keep looking for a new message.
This enables the script to start the robot, and wait for the stop-command to stop it again.
The most important feature to make this work is the sock.setblocking(0)-command on line 18, which unblocks certain commands on the sock-object, like the accept-command.